Added a "More details" button to the acts and rewards pages.
Default amount of acts is now -1(unlimited).

The amount of acts and rewards are displayed only if it's not unlimited.

Acts and rewards can now be marked as unlimited.

A new field "how_to_submit_evidences" has been added to Acts.

This new field can now be added, edited and read

User's email is displayed to reward providers when reward is requested

Updated the title and descriptions of all pages

Leaving the amount of acts or rewards empty makes it unlimited